libser4010: Try to recover from out-of-sync communication
If a command time outs because it takes very long, libser4010 will
return an error to the calling program. However if the ser4010 command
does succeed after the timeout, it will still return a result packet.
This packet is stored in the OS receive buffer. The next time a command
is send to the ser4010 device, libser4010 will directly read the
previous response from the OS buffer. libser4010 would detect this
response was for a different command, and return an error. In the mean
time the ser4010 returns a new response for the current command which in
turn is queued in the OS buffer. This situation will continue till the
serial port is closed and reopened.

Change the receiving code to ignore out-of-sync responses and continue
waiting for the correct response packet or timeout.